refac: removed unused components
